Move queryable projects into main project.
