--- URF.Core.Abstractions.Queryable/URF.Core.Abstractions.Queryable.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <LangVersion>latest</LangVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\URF.Core.Abstractions\URF.Core.Abstractions.csproj" />
-  </ItemGroup>
-
-</Project>
