--- URF.Core.EF.Queryable/URF.Core.EF.Queryable.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="2.0.1" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="2.0.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\URF.Core.Abstractions.Queryable\URF.Core.Abstractions.Queryable.csproj" />
-    <ProjectReference Include="..\URF.Core.Abstractions\URF.Core.Abstractions.csproj" />
-    <ProjectReference Include="..\URF.Core.EF\URF.Core.EF.csproj" />
-  </ItemGroup>
-
-</Project>
