+++ URF.Core.EF.Tests/URF.Core.EF.Tests.csproj
@@ -18,10 +18,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\URF.Core.Abstractions.Queryable\URF.Core.Abstractions.Queryable.csproj" />
     <ProjectReference Include="..\URF.Core.Abstractions.Trackable\URF.Core.Abstractions.Trackable.csproj" />
     <ProjectReference Include="..\URF.Core.Abstractions\URF.Core.Abstractions.csproj" />
-    <ProjectReference Include="..\URF.Core.EF.Queryable\URF.Core.EF.Queryable.csproj" />
     <ProjectReference Include="..\URF.Core.EF.Trackable\URF.Core.EF.Trackable.csproj" />
     <ProjectReference Include="..\URF.Core.EF\URF.Core.EF.csproj" />
   </ItemGroup>
