+++ URF.Core.EF.Trackable/URF.Core.EF.Trackable.csproj
@@ -13,10 +13,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\URF.Core.Abstractions.Queryable\URF.Core.Abstractions.Queryable.csproj" />
     <ProjectReference Include="..\URF.Core.Abstractions.Trackable\URF.Core.Abstractions.Trackable.csproj" />
     <ProjectReference Include="..\URF.Core.Abstractions\URF.Core.Abstractions.csproj" />
-    <ProjectReference Include="..\URF.Core.EF.Queryable\URF.Core.EF.Queryable.csproj" />
     <ProjectReference Include="..\URF.Core.EF\URF.Core.EF.csproj" />
   </ItemGroup>
